backend/requirements.txt
fastapi
uvicorn
pydantic
yfinance 
pandas
backtesting
ta
alpaca-trade-api
numpy 
quantstats
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
newsapi-python>=0.2.7
finnhub-python
autogen-agentchat
autogen-core
autogenstudio
xhtml2pdf>=0.2.11
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
# FRED API
fredapi>=0.5.0
duckduckgo_search>=5.3.0
ddgs>=3.9.0
gnews>=0.3.9
vaderSentiment>=3.3.2
# Multi-Agent Chat System
crewai>=0.28.0
groq>=0.4.0
websockets>=12.0
